fix: adapt to filebrowser code refactoring (filebrowser/filebrowser#506)
